% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempHistogram.R
\name{AntHistory}
\alias{AntHistory}
\title{Makes a histogram based on the min_ma}
\usage{
AntHistory(Ants, min_ma)
}
\arguments{
\item{Ants}{is the dataset}

\item{min_ma}{is the variable column being graphed}
}
\value{
a histogram showing how many ants lived at each minimum age that they could have been alive.
}
\description{
Makes a histogram based on the min_ma
}
