% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntsIn.R
\name{AntsIn}
\alias{AntsIn}
\title{3rd R command.
Input: The genus column of ants.
Output: The number of ants per genus.}
\usage{
AntsIn(Ants, ColumnGenus)
}
\arguments{
\item{Ants}{is the data set Ants.csv}

\item{ColumnGenus}{is the genus of the Ants, and variable column being graphed}
}
\value{
GenusCount A histogram of the minimum age of the ants.
}
\description{
3rd R command.
Input: The genus column of ants.
Output: The number of ants per genus.
}
