% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntsIn.R
\name{avg}
\alias{avg}
\title{3rd R command.
Input: splitting min and max values into averages.
Output: column of averages for each taxon}
\usage{
avg(Ants, ColumnGenus)
}
\arguments{
\item{Ants}{is the data set Ants.csv}

\item{ColumnGenus}{is the genus of the Ants}
}
\value{
GenusCount The count of ants for each genus
}
\description{
3rd R command.
Input: splitting min and max values into averages.
Output: column of averages for each taxon
}
