% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{First R script.
Input: removes NA values
Output: a data set free of missing values
Remove NA values from a data set.}
\usage{
cleanup(data)
}
\arguments{
\item{input_value}{data with NA values}
}
\value{
output_value
}
\description{
First R script.
Input: removes NA values
Output: a data set free of missing values
Remove NA values from a data set.
}
